pyproject.toml
[project]
name = "statpp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "numpy>=2.3.1",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]
